Moved serialization code to its own source files.
