+++ src/jsusfx.h
@@ -35,7 +35,6 @@ class JsusFx;
 struct JsusFxFileAPI;
 struct JsusFxGfx;
 struct JsusFxPathLibrary;
-struct JsusFxSerializationData;
 struct JsusFxSerializer;
 
 struct JsusFx_FileInfo;
@@ -231,168 +230,4 @@ public:
     
     // ==============================================================
     eel_string_context_state *m_string_context;
-};
-
-struct JsusFxSerializationData
-{
-	// note : Reaper will always save/restore values using single precision floating point, so we don't use EEL_F here,
-	// as it may be double or float depending on compile time options
-	
-	struct Slider
-	{
-		int index;
-		float value;
-	};
-	
-	std::vector<Slider> sliders;
-	std::vector<float> vars;
-	
-	void addSlider(const int index, const float value)
-	{
-		sliders.resize(sliders.size() + 1);
-		
-		Slider & slider = sliders.back();
-		slider.index = index;
-		slider.value = value;
-	}
-	
-	void addVar(const float value)
-	{
-		vars.push_back(value);
-	}
-};
-
-struct JsusFxSerializer
-{
-	virtual void begin(JsusFx & _jsusFx, const bool _write) = 0;
-	virtual void end() = 0;
-	
-	virtual int file_avail() const = 0;
-	virtual int file_var(EEL_F & value) = 0;
-	virtual int file_mem(EEL_F * values, const int numValues) = 0;
-};
-
-struct JsusFxSerializer_Basic : JsusFxSerializer
-{
-	JsusFx * jsusFx;
-	bool write;
-	
-	JsusFxSerializationData & serializationData;
-	
-	int varPosition;
-	
-	JsusFxSerializer_Basic(JsusFxSerializationData & _serializationData)
-		: jsusFx(nullptr)
-		, write(false)
-		, serializationData(_serializationData)
-		, varPosition(0)
-	{
-	}
-	
-	virtual void begin(JsusFx & _jsusFx, const bool _write) override
-	{
-		jsusFx = &_jsusFx;
-		write = _write;
-		
-		varPosition = 0;
-		
-		if (write)
-			saveSliders(*jsusFx, serializationData);
-		else
-			restoreSliders(*jsusFx, serializationData);
-	}
-	
-	virtual void end() override
-	{
-	}
-	
-	static void saveSliders(const JsusFx & jsusFx, JsusFxSerializationData & serializationData)
-	{
-		for (int i = 0; i < jsusFx.kMaxSliders; ++i)
-		{
-			if (jsusFx.sliders[i].exists)
-			{
-				serializationData.addSlider(i, jsusFx.sliders[i].getValue());
-			}
-		}
-	}
-	
-	static void restoreSliders(JsusFx & jsusFx, const JsusFxSerializationData & serializationData)
-	{
-		for (int i = 0; i < serializationData.sliders.size(); ++i)
-		{
-			const JsusFxSerializationData::Slider & slider = serializationData.sliders[i];
-			
-			if (slider.index >= 0 &&
-				slider.index < JsusFx::kMaxSliders &&
-				jsusFx.sliders[slider.index].exists)
-			{
-				jsusFx.sliders[slider.index].setValue(slider.value);
-			}
-		}
-	}
-	
-	virtual int file_avail() const override
-	{
-		if (write)
-			return -1;
-		else
-			return varPosition == serializationData.vars.size() ? 0 : 1;
-	}
-	
-	virtual int file_var(EEL_F & value) override
-	{
-		if (write)
-		{
-			serializationData.vars.push_back(value);
-			
-			return 1;
-		}
-		else
-		{
-			if (varPosition >= 0 && varPosition < serializationData.vars.size())
-			{
-				value = serializationData.vars[varPosition];
-				varPosition++;
-				return 1;
-			}
-			else
-			{
-				value = 0.f;
-				return 0;
-			}
-		}
-	}
-	
-	virtual int file_mem(EEL_F * values, const int numValues) override
-	{
-		if (write)
-		{
-			for (int i = 0; i < numValues; ++i)
-			{
-				serializationData.vars.push_back(values[i]);
-			}
-			return 1;
-		}
-		else
-		{
-			if (numValues < 0)
-				return 0;
-			if (varPosition >= 0 && varPosition + numValues <= serializationData.vars.size())
-			{
-				for (int i = 0; i < numValues; ++i)
-				{
-					values[i] = serializationData.vars[varPosition];
-					varPosition++;
-				}
-				return 1;
-			}
-			else
-			{
-				for (int i = 0; i < numValues; ++i)
-					values[i] = 0.f;
-				return 0;
-			}
-		}
-	}
 };
